% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clplot.R
\docType{package}
\name{clplot}
\alias{clplot}
\alias{clplot-package}
\title{clplot: Plotting tools for the House of Commons Library}
\description{
The clplot package provides themes and tools for making plots in the House
of Commons Library style.
}
