% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{commonslib_color}
\alias{commonslib_color}
\title{Get the hex code for a commons library color}
\usage{
commonslib_color(color_name)
}
\arguments{
\item{color_name}{The name of the color.}
}
\description{
Get the hex code for a commons library color
}
