% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{commonslib_colors}
\alias{commonslib_colors}
\title{Commons Library colors}
\format{An object of class \code{character} of length 8.}
\usage{
commonslib_colors
}
\description{
Commons Library colors
}
\keyword{datasets}
