% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{get_commonslib_green_colors}
\alias{get_commonslib_green_colors}
\title{Function to extract commonslib_green colors as hex codes}
\usage{
get_commonslib_green_colors(...)
}
\arguments{
\item{...}{Names of colors in \code{commonslib_green_colors}.}
}
\description{
Function to extract commonslib_green colors as hex codes
}
