% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_commonslib_green}
\alias{theme_commonslib_green}
\title{Commons Library Green theme}
\usage{
theme_commonslib_green(subtitle = TRUE, axes = "bl", grid = "")
}
\arguments{
\item{subtitle}{Boolean to indicate whether the plot has a subtitle. This
argument controls the spacing after the title, so that it is smaller when
a subtitle is present. The default is TRUE.}

\item{axes}{A string to indicate which axes should have axis lines and
ticks. Designate the axes to show by including a particular character in
the string: "t" for top, "r" for right, "b" for bottom, "l" for left. You
will need to position the axes correctly with ggplot, and turn on any
secondary axes, in order for the specified axes lines and ticks to be
displayed. The default is "bl", meaning both axes are shown by
default.}

\item{grid}{A string to indicate which gridlines should be shown. Designate
which gridlines to show by including a particular character in the string:
"h" for horizontal, "v" for vertical. The default is an empty string, meaning
no gridlines are shown by default.}
}
\description{
Commons Library Green theme
}
